Extract and create a ParameterBuilder to separate building from validating responsibilities. Change the namespace
