+++ lib/stack_master/sparkle_formation/compile_time/empty_validator.rb
@@ -2,7 +2,7 @@ require_relative 'validator'
 
 module StackMaster
   module SparkleFormation
-    module CompileTimeParameter
+    module CompileTime
       class EmptyValidator < Validator
 
         def initialize(name, _definition, value)
