+++ lib/stack_master/sparkle_formation/compile_time/max_length_validator.rb
@@ -2,10 +2,10 @@ require_relative 'validator'
 
 module StackMaster
   module SparkleFormation
-    module CompileTimeParameter
-      class MinSizeValidator < Validator
+    module CompileTime
+      class MaxLengthValidator < Validator
 
-        KEY = :min_size
+        KEY = :max_length
 
         def initialize(name, parameter_definition, value)
           @name = name
@@ -17,15 +17,15 @@ module StackMaster
 
         def check_is_valid
           return true unless @parameter_definition.key?(KEY)
-          !value_is_less_than_min_size?
+          !value_is_greater_than_max_length?
         end
 
-        def value_is_less_than_min_size?
-          @value.to_i < @parameter_definition[KEY].to_i
+        def value_is_greater_than_max_length?
+          @value.length > @parameter_definition[KEY].to_i
         end
 
         def create_error
-          "#{@name}:#{@value} must not be less than #{KEY}:#{@parameter_definition[KEY]}"
+          "#{@name}:#{@value} must not exceed #{KEY}:#{@parameter_definition[KEY]} characters"
         end
 
       end
