+++ lib/stack_master/sparkle_formation/compile_time/min_size_validator.rb
@@ -2,10 +2,10 @@ require_relative 'validator'
 
 module StackMaster
   module SparkleFormation
-    module CompileTimeParameter
-      class MaxSizeValidator < Validator
+    module CompileTime
+      class MinSizeValidator < Validator
 
-        KEY = :max_size
+        KEY = :min_size
 
         def initialize(name, parameter_definition, value)
           @name = name
@@ -17,15 +17,15 @@ module StackMaster
 
         def check_is_valid
           return true unless @parameter_definition.key?(KEY)
-          !value_is_greater_than_max_size?
+          !value_is_less_than_min_size?
         end
 
-        def value_is_greater_than_max_size?
-          @value.to_i > @parameter_definition[KEY].to_i
+        def value_is_less_than_min_size?
+          @value.to_i < @parameter_definition[KEY].to_i
         end
 
         def create_error
-          "#{@name}:#{@value} must not be greater than #{KEY}:#{@parameter_definition[KEY]}"
+          "#{@name}:#{@value} must not be less than #{KEY}:#{@parameter_definition[KEY]}"
         end
 
       end
